Move support utilities into `BeamUtil`
